+++ src/restorecontroller.cpp
@@ -33,6 +33,13 @@ void RestoreController::setPassword(const QString& password)
     update();
 }
 
+void RestoreController::setDefaultName(const QString &default_name)
+{
+    if (m_default_name == default_name) return;
+    m_default_name = default_name;
+    emit defaultNameChanged(m_default_name);
+}
+
 void RestoreController::setName(const QString& name)
 {
     if (m_name == name) return;
@@ -84,7 +91,11 @@ void RestoreController::accept()
 void RestoreController::update()
 {
     auto check = [this] {
-        if (!m_network) return false;
+        if (!m_network) {
+            setDefaultName("");
+            return false;
+        }
+        setDefaultName(WalletManager::instance()->newWalletName(m_network));
         if (m_mnemonic.length() == 27) {
             if (m_password.isEmpty()) return false;
         } else if (m_mnemonic.length() == 24) {